+++ RELEASE_NOTES.md
@@ -1236,3 +1236,15 @@
 ## 02/01/2021
 #### Version 13.1.0.0.20
 - [ADD] clx_retail_pricelist : set management company price list on sale order.
+
+## 02/01/2021
+#### Version 13.1.1.0.27
+- [FIX] clx_subscription_creation : deleted subscription when subscription have only when line and user delete that sale orders.
+
+## 02/01/2021
+#### Version 13.3.0.0.37
+- [FIX] clx_budget_analysis_report : set default value on budget pivot report wizard of the lasr record.
+
+## 02/01/2021
+#### Version 13.1.1.0.34
+- [FIX] contact_modification : added onchange for set customer company pricelist from the management company recors.